// src/utils/types.ts
export type TTicket = {
    origin: string;
    origin_name: string;
    destination: string;
    destination_name: string;
    departure_date: string;
    departure_time: string;
    arrival_date: string;
    arrival_time: string;
    carrier: string;
    stops: number;
    price: number;
}

export type TCurrency = "RUB" | "USD" | "EUR";

export type TFilterStops = number[] | "all";